fixes from comments: check if dump exists, remove env values

+++ publisher/start.sh
@@ -10,11 +10,18 @@ plyo/postgres:9.5.10-1.0.1
 docker exec plyo_postgres mkdir /dumps
 backup_date=`date +%Y-%m-%d-%H_00`
 backup_file="${backup_date}"-preview-db-5432-hourly.backup
-docker exec plyo_postgres cp /files/${backup_file} /dumps/plyo.backup
-docker exec plyo_postgres cp /restore.sh /docker-entrypoint-initdb.d/restore.sh
 
-docker commit $(docker ps -f name=plyo_postgres -q) plyo/postgres:publisher
-docker push plyo/postgres:publisher
+if [ $(docker exec plyo_postgres test -e "/files/$backup_file" && echo $?) ];
+then
+    docker exec plyo_postgres cp /files/${backup_file} /dumps/plyo.backup
+#    restore.sh will be executed at the moment of container's start
+    docker exec plyo_postgres cp /restore.sh /docker-entrypoint-initdb.d/restore.sh
+
+    docker commit $(docker ps -f name=plyo_postgres -q) plyo/postgres:publisher
+    docker push plyo/postgres:publisher
+else
+    echo file "${backup_file}" is not found
+fi
 
 docker kill plyo_postgres
 docker rm plyo_postgres
